--- a/db.go
+++ b/db.go
@@ -142,14 +142,17 @@ func NewPersistentDB(path string) (*DB, error) {
 //   - metadata: Optional metadata to associate with the collection.
 //   - embeddingFunc: Optional function to use to embed documents.
 //     Uses the default embedding function if not provided.
-func (db *DB) CreateCollection(name string, metadata map[string]string, embeddingFunc EmbeddingFunc) (*Collection, error) {
+//   - normalized: Optional flag to indicate if the embeddings of the collection
+//     are normalized (when you add embeddings yourself, or the embeddings created
+//     by the embeddingFunc). If nil it will be autodetected.
+func (db *DB) CreateCollection(name string, metadata map[string]string, embeddingFunc EmbeddingFunc, normalized *bool) (*Collection, error) {
 	if name == "" {
 		return nil, errors.New("collection name is empty")
 	}
 	if embeddingFunc == nil {
 		embeddingFunc = NewEmbeddingFuncDefault()
 	}
-	collection, err := newCollection(name, metadata, embeddingFunc, db.persistDirectory)
+	collection, err := newCollection(name, metadata, embeddingFunc, normalized, db.persistDirectory)
 	if err != nil {
 		return nil, fmt.Errorf("couldn't create collection: %w", err)
 	}
@@ -213,12 +216,15 @@ func (db *DB) GetCollection(name string, embeddingFunc EmbeddingFunc) *Collectio
 //   - metadata: Optional metadata to associate with the collection.
 //   - embeddingFunc: Optional function to use to embed documents.
 //     Uses the default embedding function if not provided.
-func (db *DB) GetOrCreateCollection(name string, metadata map[string]string, embeddingFunc EmbeddingFunc) (*Collection, error) {
+//   - normalized: Optional flag to indicate if the embeddings of the collection
+//     are normalized (when you add embeddings yourself, or the embeddings created
+//     by the embeddingFunc). If nil it will be autodetected.
+func (db *DB) GetOrCreateCollection(name string, metadata map[string]string, embeddingFunc EmbeddingFunc, normalized *bool) (*Collection, error) {
 	// No need to lock here, because the methods we call do that.
 	collection := db.GetCollection(name, embeddingFunc)
 	if collection == nil {
 		var err error
-		collection, err = db.CreateCollection(name, metadata, embeddingFunc)
+		collection, err = db.CreateCollection(name, metadata, embeddingFunc, normalized)
 		if err != nil {
 			return nil, fmt.Errorf("couldn't create collection: %w", err)
 		}