--- a/query.go
+++ b/query.go
@@ -96,6 +96,9 @@ func documentMatchesFilters(document *Document, where, whereDocument map[string]
 }
 
 func calcDocSimilarity(ctx context.Context, queryVectors []float32, docs []*Document) ([]*docSim, error) {
+	similarities := make([]*docSim, 0, len(docs))
+	similaritiesLock := sync.Mutex{}
+
 	// Determine concurrency. Use number of docs or CPUs, whichever is smaller.
 	numCPUs := runtime.NumCPU()
 	numDocs := len(docs)
@@ -119,7 +122,7 @@ func calcDocSimilarity(ctx context.Context, queryVectors []float32, docs []*Docu
 		}
 	}
 
-	resChan := make(chan *docSim, concurrency*2)
+	wg := sync.WaitGroup{}
 	// Instead of using a channel to pass documents into the goroutines, we just
 	// split the slice into sub-slices and pass those to the goroutines.
 	// This turned out to be faster in the query benchmarks.
@@ -133,7 +136,9 @@ func calcDocSimilarity(ctx context.Context, queryVectors []float32, docs []*Docu
 			end += rem
 		}
 
+		wg.Add(1)
 		go func(subSlice []*Document) {
+			defer wg.Done()
 			for _, doc := range subSlice {
 				// Stop work if another goroutine encountered an error.
 				if ctx.Err() != nil {
@@ -146,19 +151,18 @@ func calcDocSimilarity(ctx context.Context, queryVectors []float32, docs []*Docu
 					return
 				}
 
-				resChan <- &docSim{docID: doc.ID, similarity: sim}
+				similaritiesLock.Lock()
+				// We don't defer the unlock because we want to unlock much earlier.
+				similarities = append(similarities, &docSim{docID: doc.ID, similarity: sim})
+				similaritiesLock.Unlock()
 			}
 		}(docs[start:end])
 	}
 
-	similarities := make([]*docSim, 0, len(docs))
-	for i := 0; i < len(docs); i++ {
-		select {
-		case res := <-resChan:
-			similarities = append(similarities, res)
-		case <-ctx.Done():
-			return nil, sharedErr
-		}
+	wg.Wait()
+
+	if sharedErr != nil {
+		return nil, sharedErr
 	}
 
 	return similarities, nil