… README.md
… collection_test.go
… db_test.go
… embed_ollama_test.go
… embed_openai_test.go
… examples/rag-wikipedia-ollama/README.md
… examples/rag-wikipedia-ollama/main.go
… examples/semantic-search-arxiv-openai/README.md
… examples/semantic-search-arxiv-openai/main.go
… persistence_test.go
… vector.go